+++ kaffe/kaffevm/systems/unix-pthreads/thread-internal.h
@@ -357,6 +357,6 @@ void KaffePThread_WaitForResume(int releaseMutex, unsigned int state);
 void KaffePThread_AckAndWaitForResume(jthread_t cur, unsigned int newState);
 int KaffePThread_getSuspendSignal(void);
 
-void detectStackBoundaries(jthread_t jtid, int mainThreadStackSize);
+void detectStackBoundaries(jthread_t jtid, size_t mainThreadStackSize);
 
 #endif /* __thread_impl_h */
